Add deploy target for building all static binaries
The make deploy target invokes scripts/deploy.sh.
This script builds ckains and uses it for creating in turn
Ubuntu 12.04 x86_64 and i686 static ckains binaries.
Add deployment of both static binaries to Github in .travis.yml.
For deployment to work, the travis build was changed to
require a VM instad of a container.