Do a pivot_root() when changing to new rootfs
Change implementation to actually do a pivot_root() when
getting into new rootfs. Otherwise the old rootfs mount
is not actually moved to a location where it can be
unmounted correctly, which leads to dangling mounts
from the parent.

In particular, this fixes issues with automounted
filesystems that can't correctly expire/unmount
while a ckains process is running.